Dodata provera uloge autorizovanog korisnika prilikom pristupa funkcijama kontrolera

+++ spa-app/app/Http/Controllers/ProviderController.php
@@ -54,6 +54,9 @@ class ProviderController extends Controller
         if ($validator->fails())
             return response()->json($validator->errors());
 
+        if(auth()->user()->isUser())
+            return response()->json('You are not authorized to create new providers.');      
+
         $provider = Provider::create([
             'name' => $request->name,
             'phone_number' => $request->phone_number,
@@ -105,6 +108,9 @@ class ProviderController extends Controller
         if ($validator->fails())
             return response()->json($validator->errors());
 
+        if(auth()->user()->isUser())
+            return response()->json('You are not authorized to update providers.');      
+
         $provider->name = $request->name;
         $provider->phone_number = $request->phone_number;
         $provider->years_of_experience = $request->years_of_experience;
@@ -112,7 +118,7 @@ class ProviderController extends Controller
         
         $provider->save();
 
-        return response()->json(['Provider is created successfully.', new ProviderResource($provider)]);
+        return response()->json(['Provider is updated successfully.', new ProviderResource($provider)]);
     }
 
     /**
@@ -123,6 +129,10 @@ class ProviderController extends Controller
      */
     public function destroy(Provider $provider)
     {
+
+        if(auth()->user()->isUser())
+            return response()->json('You are not authorized to delete providers.'); 
+
         $provider->delete();
 
         return response()->json('Provider is deleted successfully.');
